requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.2
pydantic[email]==2.6.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pymongo==4.6.2
bcrypt==4.1.2
python-multipart>=0.0.9
APScheduler==3.11.2
